Rename: 컴포넌트 폴더 구분 변경
