+++ src/app/layout.tsx
@@ -1,8 +1,8 @@
 import './globals.css';
 import type { Metadata } from 'next';
 import { Open_Sans } from 'next/font/google';
-import Header from '@/components/Header';
-import Footer from '@/components/Footer';
+import Header from '@/components/home/Header';
+import Footer from '@/components/home/Footer';
 import SWRConfigContext from '@/context/SWRConfigContext';
 
 const openSans = Open_Sans({
