+++ src/components/home/PinnedPosts.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import PostCard from './PostCard';
+import PostCard from '../ui/PostCard';
 import usePosts from '@/hooks/usePosts';
 
 export default function PinnedPosts() {
@@ -11,7 +11,7 @@ export default function PinnedPosts() {
     <ul className='mx-auto grid grid-cols-1 tablet:grid-cols-2  laptop:grid-cols-3 gap-6'>
       {pinnedPosts?.map((post, index) => (
         <li key={index}>
-        <PostCard  post={post} />
+          <PostCard post={post} />
         </li>
       ))}
     </ul>
