+++ src/components/post/AdjacentPostCard.tsx
